+++ src/main/java/survivalblock/eye_of_dreams/mixin/phantomchanges/MobEntityMixin.java
@@ -6,7 +6,6 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.entity.mob.PhantomEntity;
 import net.minecraft.server.world.ServerWorld;
-import org.spongepowered.asm.mixin.Debug;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -20,8 +19,8 @@ public class MobEntityMixin {
         if (!((MobEntity) (Object) this instanceof PhantomEntity phantomEntity)) {
             return;
         }
-        phantomEntity.setPhantomSize(phantomEntity.getPhantomSize() + phantomEntity.getRandom().nextBetween(1, 3));
-        phantomEntity.heal(damage * 0.4f);
+        phantomEntity.setPhantomSize(phantomEntity.getPhantomSize() + phantomEntity.getRandom().nextBetween(1, 2));
+        phantomEntity.heal(damage * 0.2f);
         if (target instanceof LivingEntity living) {
             living.startRiding(phantomEntity);
         }