+++ src/main/java/survivalblock/eye_of_dreams/mixin/phantomchanges/PhantomEntityMixin.java
@@ -1,6 +1,5 @@
 package survivalblock.eye_of_dreams.mixin.phantomchanges;
 
-import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.minecraft.entity.EntityType;
@@ -8,7 +7,6 @@ import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.entity.mob.PhantomEntity;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
-import org.spongepowered.asm.mixin.Debug;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
@@ -19,14 +17,6 @@ public abstract class PhantomEntityMixin extends MobEntity {
         super(entityType, world);
     }
 
-    @WrapMethod(method = "tick")
-    private void noClipOn(Operation<Void> original) {
-        this.noClip = true;
-        original.call();
-        this.noClip = false;
-    }
-
-    @Debug(export = true)
     @Mixin(targets = "net.minecraft.entity.mob.PhantomEntity$PhantomMoveControl")
     public static class PhantomMoveControlMixin {
         @WrapOperation(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/util/math/Vec3d;multiply(D)Lnet/minecraft/util/math/Vec3d;"))