+++ changelog.md
@@ -1,2 +1,2 @@
-Eye of Dreams 1.0.4+1.21.8
-- Fix rare crash where a translation string was empty
+Eye of Dreams 1.0.5+1.21.8
+- Nerf phantoms, but reduce their hitbox